src/main/kotlin/no/nav/amt/person/service/nav_bruker/NavBrukerService.kt
package no.nav.amt.person.service.nav_bruker

import no.nav.amt.person.service.clients.krr.KrrProxyClient
import no.nav.amt.person.service.clients.pdl.PdlClient
import no.nav.amt.person.service.config.SecureLog.secureLog
import no.nav.amt.person.service.nav_ansatt.NavAnsatt
import no.nav.amt.person.service.nav_ansatt.NavAnsattService
import no.nav.amt.person.service.nav_enhet.NavEnhet
import no.nav.amt.person.service.nav_enhet.NavEnhetService
import no.nav.amt.person.service.person.PersonService
import no.nav.amt.person.service.person.RolleService
import no.nav.amt.person.service.person.model.Person
import no.nav.amt.person.service.person.model.Rolle
import no.nav.amt.person.service.person.model.erBeskyttet
import no.nav.amt.person.service.utils.EnvUtils
import no.nav.poao_tilgang.client.PoaoTilgangClient
import org.slf4j.LoggerFactory
import org.springframework.stereotype.Service
import java.util.*

@Service
class NavBrukerService(
	private val repository: NavBrukerRepository,
	private val personService: PersonService,
	private val navAnsattService: NavAnsattService,
	private val navEnhetService: NavEnhetService,
	private val rolleService: RolleService,
	private val krrProxyClient: KrrProxyClient,
	private val poaoTilgangClient: PoaoTilgangClient,
	private val pdlClient: PdlClient,
) {

	private val log = LoggerFactory.getLogger(javaClass)

	fun hentNavBruker(id: UUID): NavBruker {
		return repository.get(id).toModel()
	}

	fun hentNavBruker(personIdent: String): NavBruker? {
		return repository.get(personIdent)?.toModel()
	}

	fun hentEllerOpprettNavBruker(personIdent: String): NavBruker {
		return repository.get(personIdent)?.toModel() ?: opprettNavBruker(personIdent)
	}

	private fun opprettNavBruker(personIdent: String): NavBruker {
		val personOpplysninger = pdlClient.hentPerson(personIdent)

		if (personOpplysninger.adressebeskyttelseGradering.erBeskyttet()) {
			throw IllegalStateException("Nav bruker er adreessebeskyttet og kan ikke lagres")
		}

		val person = personService.hentEllerOpprettPerson(personIdent, personOpplysninger)
		val veileder = navAnsattService.hentBrukersVeileder(personIdent)
		val navEnhet = navEnhetService.hentNavEnhetForBruker(personIdent)
		val kontaktinformasjon = krrProxyClient.hentKontaktinformasjon(personIdent).getOrNull()
		val erSkjermet = poaoTilgangClient.erSkjermetPerson(personIdent).getOrThrow()

		val navBruker = NavBruker(
			id = UUID.randomUUID(),
			person = person,
			navVeileder = veileder,
			navEnhet = navEnhet,
			telefon = kontaktinformasjon?.telefonnummer ?:  personOpplysninger.telefonnummer,
			epost = kontaktinformasjon?.epost,
			erSkjermet = erSkjermet,
		)

		repository.upsert(navBruker.toUpsert())
		rolleService.opprettRolle(person.id, Rolle.NAV_BRUKER)

		return navBruker
	}

	fun oppdaterNavEnhet(navBruker: NavBruker, navEnhet: NavEnhet?) {
		repository.upsert(navBruker.toUpsert(navEnhetId = navEnhet?.id))
	}

	fun finnBrukerId(gjeldendeIdent: String): UUID? {
		return repository.finnBrukerId(gjeldendeIdent)
	}

	fun oppdaterNavVeileder(navBrukerId: UUID, veileder: NavAnsatt) {
		repository.oppdaterNavVeileder(navBrukerId, veileder.id)
	}

	fun settSkjermet(brukerId: UUID, erSkjermet: Boolean) {
		repository.settSkjermet(brukerId, erSkjermet)
	}

	fun oppdaterKontaktinformasjon(personer: List<Person>) {
		personer.forEach {person ->
			oppdaterKontaktinformasjon(person.personIdent)
		}
	}

	private fun oppdaterKontaktinformasjon(personIdent: String) {
		val eksisterendeKontaktinfo = repository.hentKontaktinformasjonHvisBrukerFinnes(personIdent) ?: return

		val krrKontaktinfo = krrProxyClient.hentKontaktinformasjon(personIdent).getOrElse {
			val feilmelding = "Klarte ikke hente kontaktinformasjon fra KRR-Proxy: ${it.message}"

			if (EnvUtils.isDev()) log.info(feilmelding)
			else log.error(feilmelding)

			return
		}

		val telefon = krrKontaktinfo.telefonnummer ?: pdlClient.hentTelefon(personIdent)

		if (eksisterendeKontaktinfo.telefon == telefon && eksisterendeKontaktinfo.epost == krrKontaktinfo.epost) return

		repository.oppdaterKontaktinformasjon(eksisterendeKontaktinfo.copy(telefon = telefon, epost = krrKontaktinfo.epost))
	}

	fun slettBrukere(personer: List<Person>) {
		personer.forEach {
			repository.deleteByPersonId(it.id)
			rolleService.fjernRolle(it.id, Rolle.NAV_BRUKER)

			if (!rolleService.harRolle(it.id, Rolle.ARRANGOR_ANSATT)) {
				personService.slettPerson(it)
			}

			secureLog.info("Slettet navbruker med personident: ${it.personIdent}")
			log.info("Slettet navbruker med personId: ${it.id}")

		}

	}

}
